# Improve performance of deduplication