The old code grows deduplication table when it is almost full. This causes a big degradation in performance for looking up non existing strings when the table is almost full because we always get hash a collision and have to iterate almost the whole table. Some tuning can be done for MaxDeduplicationTableFillFactor but 0.9 seems an ok value, not much perf improvement going lower.

This is to try and fix issues in entity service where it intermittently gets big spikes in CPU usage. 
It may also have played a part in this outage. https://confluence.derivco.co.za/display/gProducts/Entity+service+downtime+in+OSR

Small improvements in the Addition benchmark and big improvement in the new Filter benchmark:

old:
| Method   | PoolSize | Mean        | Error     | StdDev    |
|--------- |--------- |------------:|----------:|----------:|
| DoTryGet | 10000    |    13.48 ns |  0.064 ns |  0.057 ns |
| DoTryGet | 32525    | 9,420.57 ns | 91.941 ns | 76.775 ns |
| DoTryGet | 100000   |    20.26 ns |  0.351 ns |  0.329 ns |
| DoTryGet | 1000000  |   587.54 ns |  8.753 ns |  8.188 ns |

new:
| Method   | PoolSize | Mean     | Error    | StdDev   |
|--------- |--------- |---------:|---------:|---------:|
| DoTryGet | 10000    | 10.95 ns | 0.175 ns | 0.163 ns |
| DoTryGet | 32525    | 14.86 ns | 0.105 ns | 0.099 ns |
| DoTryGet | 100000   | 13.51 ns | 0.149 ns | 0.140 ns |
| DoTryGet | 1000000  | 10.87 ns | 0.124 ns | 0.110 ns |

Size 32525 is a degenerate case where the old table is almost full